requirements.txt
aiohttp==3.6.3
aiohttp-cors==0.7.0
aiohttp-rest-framework==0.0.1
aiopg==1.0.0
async-timeout==3.0.1
attrs==20.2.0
chardet==3.0.4
idna==2.10
iniconfig==1.1.1
marshmallow==3.8.0
multidict==4.7.6
packaging==20.4
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.9.0
pyparsing==2.4.7
pytest==6.1.1
pytest-aiohttp==0.3.0
PyYAML==5.3.1
six==1.15.0
SQLAlchemy==1.3.20
toml==0.10.1
yarl==1.5.1
